requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.13.0
azure-storage-blob>=12.14.0
azure-data-tables>=12.4.0
Pillow>=9.5.0
python-dotenv>=1.0.0
